test: migrate restore archived workflow run tests to testcontainers
