+++ api/tests/test_containers_integration_tests/services/test_restore_archived_workflow_run.py
@@ -2,16 +2,42 @@
 Testcontainers integration tests for workflow run restore functionality.
 """
 
+from __future__ import annotations
+
+from datetime import datetime
 from uuid import uuid4
 
 from sqlalchemy import select
 
-from models.workflow import WorkflowPause
+from models.workflow import WorkflowPause, WorkflowRun
 from services.retention.workflow_run.restore_archived_workflow_run import WorkflowRunRestore
 
 
 class TestWorkflowRunRestore:
     """Tests for the WorkflowRunRestore class."""
+
+    def test_restore_initialization(self):
+        """Restore service should respect dry_run flag."""
+        restore = WorkflowRunRestore(dry_run=True)
+
+        assert restore.dry_run is True
+
+    def test_convert_datetime_fields(self):
+        """ISO datetime strings should be converted to datetime objects."""
+        record = {
+            "id": "test-id",
+            "created_at": "2024-01-01T12:00:00",
+            "finished_at": "2024-01-01T12:05:00",
+            "name": "test",
+        }
+
+        restore = WorkflowRunRestore()
+        result = restore._convert_datetime_fields(record, WorkflowRun)
+
+        assert isinstance(result["created_at"], datetime)
+        assert result["created_at"].year == 2024
+        assert result["created_at"].month == 1
+        assert result["name"] == "test"
 
     def test_restore_table_records_returns_rowcount(self, db_session_with_containers):
         """Restore should return inserted rowcount."""
